refactor: rename pkg/papaya to pkg/core
