+++ pkg/core/.gitignore
@@ -16,4 +16,4 @@
 
 # idea
 .idea
-papaya.iml
+core.iml
