+++ pkg/core/LICENSE
