+++ pkg/core/cache/cache_test.go
