+++ pkg/core/cache/loading.go
@@ -2,7 +2,7 @@ package cache
 
 import (
 	"context"
-	"github.com/anaregdesign/lantern/pkg/papaya/model/function"
+	"github.com/anaregdesign/lantern/pkg/core/model/function"
 	"time"
 )
