+++ pkg/core/concurrent/pubsub/example/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"context"
-	"github.com/anaregdesign/lantern/pkg/papaya/concurrent/pubsub"
+	"github.com/anaregdesign/lantern/pkg/core/concurrent/pubsub"
 	"log"
 	"runtime"
 	"time"
