+++ pkg/core/concurrent/pubsub/message.go
