+++ pkg/core/concurrent/pubsub/subscription.go
@@ -2,7 +2,7 @@ package pubsub
 
 import (
 	"context"
-	"github.com/anaregdesign/lantern/pkg/papaya/model/function"
+	"github.com/anaregdesign/lantern/pkg/core/model/function"
 	"github.com/google/uuid"
 	"golang.org/x/sync/semaphore"
 	"sync"
