+++ pkg/core/graph/view.go
