+++ pkg/core/nlp/dictionary.go
