--- pkg/papaya/README.md
@@ -1,1 +1,0 @@
-# papaya
