+++ server/provider/provider.go
@@ -2,7 +2,7 @@ package provider
 
 import (
 	v1 "github.com/anaregdesign/lantern/gen/go/graph/v1"
-	"github.com/anaregdesign/lantern/pkg/papaya/cache/graph"
+	"github.com/anaregdesign/lantern/pkg/core/cache/graph"
 	"google.golang.org/grpc"
 	"net"
 	"os"
